Add files via upload
flushed out some more of the C++ backend, coming to lab with more questions on getting my input from my GUI to save in the C++. Added a few modifications to background image as well as (drumroll please)  a settings page to select between desktop and mobile view. Added a button to move from settings to info pane.